% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_installed.R
\name{assert_installed}
\alias{assert_installed}
\title{Require Packages to be Installed}
\usage{
assert_installed(pkgs)
}
\arguments{
\item{pkgs}{\verb{[character]} Packages to check for installation status}
}
\value{
invisible \code{NULL}; errors if any \code{pkgs} are not installed
}
\description{
\code{assert_installed()} is essentially \code{requireNamespace(quietly = FALSE)},
but it can handle multiple packages as a vector and outputs more helpful
error messages.
}
