% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{Create a new R Package}
\usage{
create_package(
  path,
  license = "mit",
  renv = TRUE,
  github = TRUE,
  pkgdown = github,
  coc = github,
  fledge = FALSE,
  cran_badge = fledge,
  open = interactive(),
  creator = NULL,
  ...
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{license}{\verb{[string]} The license to use; corresponds to one of the
\verb{usethis::use_*_license()} functions}

\item{renv}{\verb{[flag]} Should \code{{renv}} be initialized for this package? This
ensures that package dependencies are portable and reproducible and is
recommended.}

\item{github}{\verb{[flag]} Should a Github repository be initialized for this
package? The default is \code{TRUE} but will only work if you have set up your
environment to work with \code{usethis::use_github()}.}

\item{pkgdown}{\verb{[flag]} Should a \code{{pkgdown}} site be created for this
package? Defaults to the value of \code{github}, so that publicly available
packages get a site by default.}

\item{coc}{\verb{[flag]} Should a Code of Conduct be added to the package?
Defaults to the value of \code{github}, so that publicly available packages get
a Code of Conduct by default.}

\item{fledge}{\verb{[flag]} Should \code{{fledge}} be used for package NEWS and
versions? The default is \code{FALSE}, since I only use NEWS and version tags
for packages that I plan to distribute.}

\item{cran_badge}{\verb{[flag]} Should a badge indicating CRAN status be added to
the README? Defaults to the value of \code{fledge}, since both are mainly used
for packages that I plan to distribute.}

\item{open}{\verb{[flag]} Should the package open in a new RStudio window once
setup is complete?}

\item{creator}{\verb{[string]} The name of the package creator (aka you). This
is used to fill in the Authors field in the package Description, as well
as the contact for the Code of Conduct (if used) and the copyright holder
of the license (if applicable).}

\item{...}{Additional arguments to pass to the \code{license} function}
}
\value{
invisible \code{NULL}
}
\description{
\code{create_package()} mimics \code{usethis::create_package()}, but uses defaults that
I find more helpful. Specific actions can be found in the Details.
}
