% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithr-package.R
\docType{package}
\name{smithr-package}
\alias{smithr}
\alias{smithr-package}
\title{smithr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jesse-smith/smithr}
  \item \url{https://jesse-smith.github.io/smithr/}
  \item Report bugs at \url{https://github.com/jesse-smith/smithr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
