% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithr-package.R
\docType{package}
\name{smithr-package}
\alias{smithr}
\alias{smithr-package}
\title{smithr: A Collection of Useful Functions}
\description{
{smithr} is a collection of functions that I or others have found useful over time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jesse-smith/smithr}
  \item \url{https://jesse-smith.github.io/smithr/}
  \item Report bugs at \url{https://github.com/jesse-smith/smithr/issues}
}

}
\author{
\strong{Maintainer}: Jesse Smith \email{jesse96smith@gmail.com}

}
\keyword{internal}
